# docs/swagger.yaml
definitions:
  assignments.createAssignmentRequest:
    properties:
      created_by:
        type: string
      organization_id:
        type: string
      plan_id_or_slug:
        type: string
      user_id:
        type: string
      valid_from:
        type: string
      valid_until:
        type: string
    required:
    - created_by
    - plan_id_or_slug
    - valid_from
    type: object
  assignments.terminatePlanRequest:
    properties:
      organization_id:
        type: string
      plan_id_or_slug:
        type: string
      user_id:
        type: string
    required:
    - plan_id_or_slug
    type: object
  assignments.updateAssignedPlanRequest:
    properties:
      organization_id:
        type: string
      plan_id_or_slug:
        type: string
      set_valid_until_to_zero:
        type: boolean
      updated_by:
        type: string
      user_id:
        type: string
      valid_from:
        type: string
      valid_until:
        type: string
    required:
    - plan_id_or_slug
    - updated_by
    type: object
  errors.ErrorResponse:
    properties:
      code:
        type: string
      details:
        additionalProperties: {}
        type: object
      message:
        type: string
    type: object
...
  features.createFeatureRequest:
    properties:
      config:
        additionalProperties: {}
        type: object
      created_by:
        type: string
      description:
        type: string
      name:
        type: string
      slug:
        type: string
      type:
        enum:
        - static
        - metered
        type: string
    required:
    - created_by
    - name
    - slug
    - type
    type: object
...
  models.Feature:
    properties:
      config:
        additionalProperties: {}
        type: object
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      slug:
        type: string
      tenant_slug:
        type: string
      type:
        $ref: '#/definitions/models.FeatureTypeEnum'
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.FeatureTypeEnum:
    enum:
    - static
    - metered
    type: string
    x-enum-varnames:
    - FeatureTypeStatic
    - FeatureTypeMetered
  models.HttpResponse-array_models_Feature:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Feature'
        type: array
      message:
        type: string
      status:
        type: integer
    type: object
...
  models.HttpResponse-models_Feature:
    properties:
      data:
        $ref: '#/definitions/models.Feature'
      message:
        type: string
      status:
        type: integer
    type: object
...
  models.HttpResponse-models_PlanAssignment:
    properties:
      data:
        $ref: '#/definitions/models.PlanAssignment'
      message:
        type: string
      status:
        type: integer
    type: object
...
  models.HttpResponse-pagination_PaginationView-models_PlanAssignment:
    properties:
      data:
        $ref: '#/definitions/pagination.PaginationView-models_PlanAssignment'
      message:
        type: string
      status:
        type: integer
    type: object
  models.HttpResponse-pagination_PaginationView-models_PlanAssignmentHistory:
    properties:
      data:
        $ref: '#/definitions/pagination.PaginationView-models_PlanAssignmentHistory'
      message:
        type: string
      status:
        type: integer
    type: object
...
  models.PlanAssignment:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      organization_id:
        type: string
      plan_id:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      user_id:
        type: string
      valid_from:
        type: string
      valid_until:
        type: string
    type: object
  models.PlanAssignmentHistory:
    properties:
      action:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      organization_id:
        type: string
      plan_id:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      user_id:
        type: string
      valid_from:
        type: string
      valid_until:
        type: string
    type: object
...
  pagination.PaginationView-models_PlanAssignment:
    properties:
      limit:
        type: integer
      page:
        type: integer
      results:
        items:
          $ref: '#/definitions/models.PlanAssignment'
        type: array
      total:
        type: integer
    type: object
  pagination.PaginationView-models_PlanAssignmentHistory:
    properties:
      limit:
        type: integer
      page:
        type: integer
      results:
        items:
          $ref: '#/definitions/models.PlanAssignmentHistory'
        type: array
      total:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /v1/features:
    get:
      consumes:
      - application/json
      description: Get a list of features for the tenant
      parameters:
      - description: Tenant Slug
        in: header
        name: X-Tenant-Slug
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      - description: Filter by feature type (static/metered)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Features retrieved successfully
          schema:
            $ref: '#/definitions/models.HttpResponse-array_models_Feature'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: List features
      tags:
      - features
    post:
      consumes:
      - application/json
      description: Create a new feature for the tenant
      parameters:
      - description: Tenant Slug
        in: header
        name: X-Tenant-Slug
        required: true
        type: string
      - description: Feature creation data
        in: body
        name: feature
        required: true
        schema:
          $ref: '#/definitions/features.createFeatureRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Feature created successfully
          schema:
            $ref: '#/definitions/models.HttpResponse-models_Feature'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Create a new feature
      tags:
      - features
...
  /v1/plans/assignments:
    delete:
      consumes:
      - application/json
      description: Terminate a plan assignment for an organization or user
      parameters:
      - description: Tenant Slug
        in: header
        name: X-Tenant-Slug
        required: true
        type: string
      - description: Assignment termination information
        in: body
        name: assignment
        required: true
        schema:
          $ref: '#/definitions/assignments.terminatePlanRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Plan assignment not found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Delete a plan assignment
      tags:
      - plan-assignments
    get:
      consumes:
      - application/json
      description: Get a list of plan assignments with optional filtering
      parameters:
      - description: Tenant Slug
        in: header
        name: X-Tenant-Slug
        required: true
        type: string
      - description: 'Valid from date (format: YYYY-MM-DDThh:mm:ssZ)'
        in: query
        name: validFrom
        type: string
      - description: 'Valid until date (format: YYYY-MM-DDThh:mm:ssZ)'
        in: query
        name: validUntil
        type: string
      - description: Plan ID or slug
        in: query
        name: planIdOrSlug
        type: string
      - description: Organization ID
        in: query
        name: orgId
        type: string
      - description: User ID
        in: query
        name: userId
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Assignments retrieved successfully
          schema:
            $ref: '#/definitions/models.HttpResponse-pagination_PaginationView-models_PlanAssignment'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: List plan assignments
      tags:
      - plan-assignments
    post:
      consumes:
      - application/json
      description: Assign a plan to either an organization or a user
      parameters:
      - description: Tenant Slug
        in: header
        name: X-Tenant-Slug
        required: true
        type: string
      - description: Assignment information
        in: body
        name: assignment
        required: true
        schema:
          $ref: '#/definitions/assignments.createAssignmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Plan assignment created successfully
          schema:
            $ref: '#/definitions/models.HttpResponse-models_PlanAssignment'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Plan not found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Create a new plan assignment
      tags:
      - plan-assignments
    put:
      consumes:
      - application/json
      description: Update the validity period for a plan assignment
      parameters:
      - description: Tenant Slug
        in: header
        name: X-Tenant-Slug
        required: true
        type: string
      - description: Assignment update information
        in: body
        name: assignment
        required: true
        schema:
          $ref: '#/definitions/assignments.updateAssignedPlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Assignment updated successfully
          schema:
            $ref: '#/definitions/models.HttpResponse-models_PlanAssignment'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Plan assignment not found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Update a plan assignment
      tags:
      - plan-assignments
  /v1/plans/assignments/history:
    get:
      consumes:
      - application/json
      description: Get historical records of plan assignments with optional filtering
      parameters:
      - description: Tenant Slug
        in: header
        name: X-Tenant-Slug
        required: true
        type: string
      - description: 'Valid from before date (format: YYYY-MM-DDThh:mm:ssZ)'
        in: query
        name: validFromBefore
        type: string
      - description: 'Valid from after date (format: YYYY-MM-DDThh:mm:ssZ)'
        in: query
        name: validFromAfter
        type: string
      - description: 'Valid until before date (format: YYYY-MM-DDThh:mm:ssZ)'
        in: query
        name: validUntilBefore
        type: string
      - description: 'Valid until after date (format: YYYY-MM-DDThh:mm:ssZ)'
        in: query
        name: validUntilAfter
        type: string
      - description: Action type (Create, Update, Delete)
        in: query
        name: action
        type: string
      - description: Plan ID or slug
        in: query
        name: planIdOrSlug
        type: string
      - description: Organization ID
        in: query
        name: orgId
        type: string
      - description: User ID
        in: query
        name: userId
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Assignment history retrieved successfully
          schema:
            $ref: '#/definitions/models.HttpResponse-pagination_PaginationView-models_PlanAssignmentHistory'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: List plan assignment history
      tags:
      - plan-assignments
swagger: "2.0"